+++ src/main/java/kh/spring/service/MealService.java
@@ -98,124 +98,7 @@ public class MealService {
 		
 		return list;
 	}
-	
-	// db저장되어있는 식단 엑셀 다운로드
-	public void excelDownload(String month, String school, HttpServletResponse response) throws IOException {		
-		Workbook wb = new XSSFWorkbook();
-		Sheet sheet = wb.createSheet(month + "월 식단표");
-		Row row = null;
-		Cell cell = null;
-		int rowNum = 0;
-		
-		// 공통 스타일
-		CellStyle commonStyle = wb.createCellStyle(); // title 셀 스타일 설정
-		commonStyle.setAlignment(HorizontalAlignment.CENTER); // 가운데 정렬
-		commonStyle.setVerticalAlignment(VerticalAlignment.CENTER);
 
-		// title
-		Font font = wb.createFont();
-		font.setBold(true); // 굵은 글씨
-		font.setFontHeightInPoints((short)14);
-		
-		CellStyle titleStyle = wb.createCellStyle(); // title 셀 스타일 설정
-		titleStyle.setAlignment(HorizontalAlignment.CENTER); // 가운데 정렬
-		titleStyle.setVerticalAlignment(VerticalAlignment.CENTER);
-		titleStyle.setFont(font);
-		
-		sheet.addMergedRegion(new CellRangeAddress(0, 1, 0, 6)); // 셀병함
-		
-		row = sheet.createRow(rowNum++);
-		cell = row.createCell(0);
-		cell.setCellStyle(titleStyle); // title 셀 스타일 적용
-		cell.setCellValue(month + "월 " + school + " 식단표");
-		
-		// 빈행 추가
-		sheet.createRow(rowNum++); 
-		row = sheet.createRow(rowNum++); 
-		
-		// Header
-		row = sheet.createRow(rowNum++);
-		cell = row.createCell(0);
-		cell.setCellStyle(commonStyle);
-		cell.setCellValue("날짜");
-		cell = row.createCell(1);
-		cell.setCellStyle(commonStyle);
-		cell.setCellValue("메뉴1");		
-		cell = row.createCell(2);
-		cell.setCellStyle(commonStyle);
-		cell.setCellValue("메뉴2");		
-		cell = row.createCell(3);
-		cell.setCellStyle(commonStyle);
-		cell.setCellValue("메뉴3");		
-		cell = row.createCell(4);
-		cell.setCellStyle(commonStyle);
-		cell.setCellValue("메뉴4");		
-		cell = row.createCell(5);
-		cell.setCellStyle(commonStyle);
-		cell.setCellValue("메뉴5");		
-		cell = row.createCell(6);
-		cell.setCellStyle(commonStyle);
-		cell.setCellValue("메뉴6");
-
-		// Body
-		CellStyle dateStyle = wb.createCellStyle(); // 날짜 셀 스타일 설정
-		dateStyle.setDataFormat(wb.getCreationHelper().createDataFormat().getFormat("yy/mm/dd (aaa)")); // 수요일 하고싶으면 aaaa 
-		dateStyle.setAlignment(HorizontalAlignment.CENTER); // 가운데 정렬
-		dateStyle.setVerticalAlignment(VerticalAlignment.CENTER);
-		
-		Map<String, String> param = new HashMap<>();
-		param.put("month", month);
-		param.put("school", school);
-
-		List<MealDTO> list = dao.excelDownloadList(param);
-		
-		for (MealDTO dto : list) {
-			row = sheet.createRow(rowNum++);
-			
-			cell = row.createCell(0);
-			cell.setCellStyle(dateStyle);
-			cell.setCellValue(dto.getMeal_date());
-			
-			cell = row.createCell(1);
-			cell.setCellStyle(commonStyle);
-			cell.setCellValue(dto.getMenu1());
-			cell = row.createCell(2);
-			cell.setCellStyle(commonStyle);
-			cell.setCellValue(dto.getMenu2());
-			cell = row.createCell(3);
-			cell.setCellStyle(commonStyle);
-			cell.setCellValue(dto.getMenu3());
-			cell = row.createCell(4);
-			cell.setCellStyle(commonStyle);
-			cell.setCellValue(dto.getMenu4());
-			cell = row.createCell(5);
-			cell.setCellStyle(commonStyle);
-			cell.setCellValue(dto.getMenu5());
-			cell = row.createCell(6);
-			cell.setCellStyle(commonStyle);
-			cell.setCellValue(dto.getMenu6());
-		}
-		
-		sheet.setColumnWidth(0, 4000);
-		for (int i=1;i<7;i++) {
-			sheet.autoSizeColumn(i);
-			sheet.setColumnWidth(i, (sheet.getColumnWidth(i)) + (short)1024); 
-			//이건 자동으로 조절 하면 너무 딱딱해 보여서 자동조정한 사이즈에 (short)512를 추가해 주니 한결 보기 나아졌다.
-		}
-			
-		// 컨텐츠 타입과 파일명 지정
-		response.setContentType("ms-vnd/excel");
-		response.setHeader("Content-Disposition", "attachment;filename=" + month + "월 " + school + " 식단표.xlsx");
-
-		// Excel File Output
-		wb.write(response.getOutputStream());
-		wb.close();
-	}
-	
-	// 엑셀에 저장되어 있는 식단 db에 업로드
-	
-	
-	
 	// 검색
 	public List<MealDTO> search(String keyword) {
 		return dao.search(keyword);
